requirements.txt
anyio==4.8.0
apiron==8.0.0.post1
asttokens==3.0.0
build==1.2.2.post1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.2.0
fonttools==4.55.6
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
matplotlib==3.10.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.2
orthanc-rest-client==0.5.7
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
pip-tools==7.4.1
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
pure_eval==0.2.3
pydicom==2.4.4
Pygments==2.19.1
pyorthanc==1.19.1
pyparsing==3.2.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
pyzmq==26.2.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
wget==3.2
